Add files via upload
This is the version 2 of my todo list terminal project and this time I used Flask. 
I was able to use the same code, installed Flask, added render html, still the same .txt file for storage of data, I also added routes for GET and POST.
Edit and Delete still functioning the same as the previous version in terminal run.